+++ src/RoutesHostServer/Models/Route.cs
@@ -14,6 +14,8 @@ namespace RoutesHostServer.Models
 		public string ApiKey { get; set; }
 		public int Priority { get; set; }
 		public string PingPath { get; set; }
+		public string MachineName { get; set; }
+		public string Ip { get; set; }
 
 		public object Clone()
 		{